Add files via upload
Data is freely available on Kaggle.com named car logo. LogoD detects logo in the image and learns unique features to each class via a neural network. It creates a .h5 extension file which is the model created by the neural network. LogoR the loads the same model file an uses it to recognize the car logo from image.